Rename views to ui
