+++ src/Instance.ts
@@ -2,8 +2,8 @@ import * as vscode from "vscode";
 import { ConnectionData, IBMiEvent } from "./typings";
 import IBMi, { ConnectionResult } from "./api/IBMi";
 import { CodeForIStorage, ConnectionStorage } from "./api/configuration/Storage";
-import { withContext } from "./views/tools";
-import { handleConnectionResults, messageCallback } from "./views/connection";
+import { withContext } from "./ui/tools";
+import { handleConnectionResults, messageCallback } from "./ui/connection";
 import { VsStorage } from "./config/Storage";
 import { VsCodeConfig } from "./config/Configuration";
 import { ConnectionConfig } from "./api/configuration/ConnectionManager";
