+++ src/commands/actions.ts
@@ -1,11 +1,11 @@
 import path from "path";
 import { commands, TreeItem, Uri, WorkspaceFolder, window, Disposable } from "vscode";
-import { refreshDiagnosticsFromServer } from "../views/diagnostics";
+import { refreshDiagnosticsFromServer } from "../ui/diagnostics";
 import Instance from "../Instance";
 import { Action, DeploymentMethod } from "../typings";
-import { runAction } from "../views/actions";
+import { runAction } from "../ui/actions";
 import IBMi from "../api/IBMi";
-import { BrowserItem } from "../views/types";
+import { BrowserItem } from "../ui/types";
 
 export function registerActionsCommands(instance: Instance): Disposable[] {
   return [
