+++ src/debug/index.ts
@@ -11,7 +11,7 @@ import { Env, getEnvConfig } from "../filesystems/local/env";
 import * as certificates from "./certificates";
 import { DEBUG_CONFIG_FILE, DebugConfiguration, getDebugServiceDetails, resetDebugServiceDetails } from "../api/configuration/DebugConfiguration";
 import * as server from "./server";
-import { withContext } from "../views/tools";
+import { withContext } from "../ui/tools";
 import { getStoredPassword } from "../config/passwords";
 
 const debugExtensionId = `IBM.ibmidebug`;
