+++ src/instantiate.ts
@@ -2,7 +2,7 @@
 import * as vscode from "vscode";
 import { onCodeForIBMiConfigurationChange } from "./config/Configuration";
 import Instance from "./Instance";
-import { Terminal } from './views/Terminal';
+import { Terminal } from './ui/Terminal';
 import { getDebugServiceDetails } from './api/configuration/DebugConfiguration';
 import { debugPTFInstalled, isDebugEngineRunning } from './debug/server';
 import { setupGitEventHandler } from './filesystems/local/git';
