+++ src/ui/diagnostics.ts
