+++ src/ui/views/LibraryListView.ts
@@ -1,9 +1,9 @@
 import vscode, { commands, l10n } from "vscode";
-import { instance } from "../instantiate";
-import { IBMiObject, WithLibrary } from "../typings";
-import { objectToToolTip } from "./tools";
-import { ConnectionConfig } from "../api/configuration/ConnectionManager";
-import IBMi from "../api/IBMi";
+import { instance } from "../../instantiate";
+import { IBMiObject, WithLibrary } from "../../typings";
+import { objectToToolTip } from "../tools";
+import { ConnectionConfig } from "../../api/configuration/ConnectionManager";
+import IBMi from "../../api/IBMi";
 
 export class LibraryListProvider implements vscode.TreeDataProvider<LibraryListNode> {
   private readonly _emitter: vscode.EventEmitter<LibraryListNode | undefined | null | void> = new vscode.EventEmitter();
